// web/src/api/generated/AccountView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AccountId } from "./AccountId";

/**
 * `GET /api/v1/account`.
 */
export type AccountView = { 
/**
 * The account id.
 */
id: AccountId, 
/**
 * The verified address.
 */
email: string, 
/**
 * When the account was created.
 */
createdAt: string, };